+++ common/src/main/java/net/frozenblock/trailiertales/registry/TTAttachmentTypes.java
@@ -17,11 +17,15 @@
 
 package net.frozenblock.trailiertales.registry;
 
+import com.mojang.serialization.Codec;
 import net.frozenblock.lib.platform.api.attachment.DataAttachmentSyncPredicate;
 import net.frozenblock.lib.platform.api.attachment.DataAttachmentType;
 import net.frozenblock.trailiertales.TTConstants;
 import net.frozenblock.trailiertales.block.entity.coffin.impl.EntityCoffinData;
+import net.frozenblock.trailiertales.block.impl.BrushableBlockAnimationState;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.storage.loot.LootTable;
 
 public final class TTAttachmentTypes {
 	public static final DataAttachmentType<ItemStack> BOAT_BANNER = DataAttachmentType.create(
@@ -34,15 +38,21 @@ public final class TTAttachmentTypes {
 	);
 	public static final DataAttachmentType<EntityCoffinData> ENTITY_COFFIN_DATA = DataAttachmentType.create(
 		TTConstants.id("entity_coffin_data"),
-		builder -> {
-			builder.persistent(EntityCoffinData.CODEC);
-		}
+		builder -> builder.persistent(EntityCoffinData.CODEC)
 	);
-	public static final DataAttachmentType<ItemStack> FALLING_BLOCK_ITEM = DataAttachmentType.create(
-		TTConstants.id("falling_block_item"),
-		builder -> {
-			builder.persistent(ItemStack.CODEC);
-		}
+
+	// BRUSHABLE BLOCK
+	public static final DataAttachmentType<BrushableBlockAnimationState> BRUSHABLE_BLOCK_ANIMATION_STATE = DataAttachmentType.create(
+		TTConstants.id("brushable_block_animation_state"),
+		builder -> builder.syncWith(BrushableBlockAnimationState.STREAM_CODEC, DataAttachmentSyncPredicate.all())
+	);
+	public static final DataAttachmentType<Boolean> BRUSHABLE_BLOCK_REBRUSHED = DataAttachmentType.create(
+		TTConstants.id("brushable_block_rebrushed"),
+		builder -> builder.persistent(Codec.BOOL)
+	);
+	public static final DataAttachmentType<ResourceKey<LootTable>> BRUSHABLE_BLOCK_STORED_LOOT_TABLE = DataAttachmentType.create(
+		TTConstants.id("brushable_block_stored_loot_table"),
+		builder -> builder.persistent(LootTable.KEY_CODEC)
 	);
 
 	public static void init() {}